Batch of CodeQL Fixes
Includes changes across the repo for the following CodeQL rules:

- cpp/comparison-with-wider-type
- cpp/overflow-buffer
- cpp/redundant-null-check-param
- cpp/uselesstest

Some false positives are filtered in CodeQlFilters.yml.

+++ ShellPkg/Library/UefiShellDebug1CommandsLib/Comp.c
@@ -367,7 +367,7 @@ ShellCommandRunComp (
       }
 
       if (ShellStatus == SHELL_SUCCESS) {
-        while (DiffPointNumber < DifferentCount) {
+        while ((UINT64)DiffPointNumber < DifferentCount) {
           DataSizeFromFile1 = 1;
           DataSizeFromFile2 = 1;
           OneByteFromFile1  = 0;
